[CT-2-2-3] fix: NumberFormatException 으로 발생하는 크래시 수정

+++ presentation/src/main/java/com/yessorae/presentation/ui/chartgame/ChartGameViewModel.kt
@@ -232,7 +232,7 @@ class ChartGameViewModel @Inject constructor(
                     val newStockCount = maxAvailableStockCount * percentage
                     old.copy(
                         tradeOrderUi = old.tradeOrderUi.copy(
-                            stockCountInput = newStockCount.toString()
+                            stockCountInput = newStockCount.toInt().toString()
                         )
                     )
                 }
